Reactivity and control flow improvements (#55)
This PR fixes many issues related to reactivity, such as cleanup
functions not being called, effects not being observed, and so forth. It
also introduces an element cache during the rendering phase, which will
reuse elements if they have not changed on a reactive re-render.

This foundation enables mapJoin to become lazy, where adding a new
element to the array being mapped will not cause previous items to
re-render. Currently, this is only true when appending items - any items
after a changed element will be re-rendered. In the future this can be
improved. In order to support this, **`mapJoin` takes a significant
breaking change** - the array or map to be mapped must be wrapped in a
callback - `mapJoin(someArray, ...)` becomes `mapJoin(() => someArray,
...)`.

Additionally, this PR adds a `For` component which allows for using
mapJoin ergonomically within the context of a JSX template. It also adds
a `Switch` component for conditionally rendering elements. Combined,
these enable moving significant amounts of logic into the template in a
way that is still readable, and without requiring any complex gymnastics
to leverage reactivity.