Add formatting and a bunch of new components (#56)
This PR implements formatting by allowing templates to include
formatting commands from Prettier. For example, a template can include
`<group>` which is a set of strings to try to keep on one line if
possible, and various kinds of breaks that say how the group should
break when it doesn't fit.

To support this, a number of additions and changes (many of them
breaking!) have been implemented:

## JSX Transform

* The JSX transform now strips all leading and trailing whitespace in
text nodes. Additionally, all spaces and line breaks in the middle of a
text node are replaced with a space. Auto-indenting behavior and the
notion of "significant" linebreaks are removed. Indents are now explicit
either via the `indent` intrinsic or `Indent` component. All significant
linebreaks must now be added via `<hbr />`, `<sbr />`, `<lbr />`, or
`<br />`.
* As a result of this change, the prettier plugin is no longer needed,
the standard JSX formatter works fine (though could be optimized to
avoid some annoying patterns).

## Alloy core

* Various listing elements get a set of props for controlling the list
behavior - `joiner`, `ender`, and boolean props for common joiner
punctuation (`comma`, `semicolon`) and line breaks (`hard`, `soft`).
* The `For` component skips falsy values.
* The `Indent` component now doesn't include an ending line break by
default, though it may be requested with `trailingBreak`. Use
`trailingBreak` when you are indenting a block with an ending sigil
(e.g. `}`), otherwise blocks should not end with a line break to allow
subsequent blocks to set the indent level for themselves).
* Added `List` component which allows joining its child elements with a
joiner/ender.
* The `Output` component now takes `printWidth`, `useTabs`, and
`tabWidth` options for controlling the default formatting of all
outputs. **The `indent` prop has been removed.**
* Added `Show` component for conditionally rendering some block of code
(similar to the `someCondition && ...` pattern, but can look nicer in
some cases.
* The `SourceFile` component now takes `printWidth`, `useTabs`, and
`tabWidth` options for controlling the default formatting for that
specific source file.
* Added `Wrap` component for conditionally wrapping a component with
other components.
* **Removed `indent` context**. It is no longer needed with indenting
handled entirely by the printer.
* The JSX runtime now supports creating intrinsic elements. Formatting
commands are created via intrinsic elements.
* Implemented `splitProps` to split a props object into two or more
objects while handling reactivity. Useful for plumbing subsets of props
into different child components.
* The renderer added the notion of a new type of Render Text Tree node
which has a `print` callback that is called by the printer to get the
rendered text for that node. Currently used for implementing formatting
for intrinsic elements.
* Added `sti` function as peer for `stc` to create intrinsic elements.
Probably not needed to be called by users as we should export all the
intrinsic elements from `@alloy-js/core/stc`.
* Added `preserveFragments` option to `children` and `childrenArray`,
which allows for using fragments in children as a list item container.

## JSON

* Now emits pretty code.

## TypeScript 

* Now emits pretty code
* Added `CommaList` component for creating lists separated by commas and
soft or hard breaks (useful for object literals, arrays and such).
* Added `JSDoc`, `JSDocComment`, `JSDocParagraph`, and `JSDocExample`
components for emitting nicely formatted JSDoc comments.
* Added `StatementList` component for creating lists separated by
semicolons and hard breaks.