Type references (#124)
fix #120

when using a typescript reference can specify it is used in a type
context which updates the imports to include `type` keyword.

If a symbol is used both as type and non type the keyword will be
omitted.
If all symbols from an import are types it uses this syntax `import type
{A, B, C} from "..."` otherwise it use `import {type A, B, type C} from
"..."`

---------

Co-authored-by: Brian Terlson <brian.terlson@microsoft.com>