[TypeScript] Fix FunctionParameters component adds an extra comma when no params and function name is long (#208)
When the function name is long (and/or deeply indented), the formatter
inserts a line break inside the function parameters parens and an extra
comma is added by `<ifBreak>,</ifBreak>`. However, when the function
have no parameters at all, it produces an invalid code snippet as
follows:

```ts
function thisFunctionNameIsTooLongSoTheFormatterWillInsertLineBreakAndIfBreakNodes(
  ,
) {}
```

This pull request fixes the component to return just `null` when no
parameters provided, so the extra comma will never be inserted by the
formatter.

---------

Co-authored-by: Timothee Guerin <timothee.guerin@outlook.com>