Refactor symbols to simplify language implementations (#206)
This PR significantly refactors symbols and introduces a number of
breaking changes for language implementations (but hopefully none or
very small for emitter packages). The goal of this PR is to make symbols
far more ergonomic and useful for languages to use and customize for
their purposes. There are also a number of related cleanups and
refactors.

## Changes to Core

* Introduce the notion of declaration and member spaces, which are keyed
containers of symbols. For example, TypeScript has "types" and "values"
spaces for lexical scopes.
* `OutputScope` and `OutputSymbol` are now abstract classes which don't
have any in-built understanding of which declaration spaces and member
spaces they have, but has a protocol for subclasses to define them. This
allows the base type to reason about any spaces that subtypes define,
enabling more work to happen in the base type and also allowing for a
centralized tracing/debugging story.
* `OutputScope` can be a member scope, which is a kind of scope that has
members which are provided by its `ownerSymbol`. This type of scope is
handled properly during refkey resolution.
* Added a `BasicScope` and `BasicSymbol` type which have the same
functionality as the old `OutputScope` and `OutputSymbol`. The built-in
scope and symbol creation components (`Scope`, `Declaration`, etc.) will
create `BasicScope`s and `BasicSymbol`s as appropriate (and complain if
you try to put one of these in a specialized scope).
* `OutputSymbol` has an abstract `copy()` method that must be
implemented by subtypes to copy the symbol properly. Today there are
bugs where e.g. an instantiated TSOutputSymbol doesn't preserve its
flags. Additionally, copy facilitates copying properties reactively, so
changes to the source will be reflected in the target.
* Flags have been removed. Flags didn't play well with reactivity (you
would get notified of any flag changing even if not the one you were
interested in), and were difficult to subclass. Nobody liked the `flags`
and `tsFlags` situation.
* Member scopes are now real scopes. The previous situation was
confusing, and treating them as real scopes allows handling member
scopes uniformly in the binder.
* Refkey resolution contains additional metadata that will help
languages build references (e.g. full path of reference and target).
* Resolution result's `targetDeclaration` has been renamed `symbol`, and
is always the exact symbol that was resolved. An additional property
`lexicalDeclaration` has been added which is the top-level lexical
declaration that contains the members specified in the member path
* Removed functionality around slots and FQN resolution. It wasn't being
used as far as I know, and was really just an experiment and not very
great anyway.
* Symbol flow is unified on the `to` direction (e.g. `moveSymbolsFrom`
is renamed to `moveSymbolsTo`). It is generalized to work on arbitrary
spaces, e.g. instantiation now requires the space key to instantiate
from and to.

## Changes to CSharp

* Removed all instances of default name-based refkeys. These are just
bad and nobody should use them. Likely only used in tests.
* Refactored more components into the new component structure.
* Improved namespace handling:
* Source files which don't have a parent `<Namespace>` are in the global
namespace.
* Source files which do have a parent namespace establish that namespace
via blockless namespace statements.
* `<Namespace>` creates block namespaces when placed inside a source
file.
* Namespace and source file symbols and scopes track C# semantics
(top-level Source File scope, nested scopes for each namespace,
declarations in namespaces create members on the owner symbol, etc.)
* `Using` now takes a namespace symbol instead of strings. `Using`s are
automatically added when needed.
* Add `LexicalScope`, `MemberScope`, and `NamespaceScope` components
which create the CSharp specific scopes.
* Add `NamespaceScopes` component to create the nested scopes for a
nested namespace symbol.
* Added CSharp scopes:
  * CSharp - the base of all scopes
* Class scope - has a place for parameters (presently just members, but
will eventually be a separate declaration space).
  * Lexical scope - has a place for local declarations
* Method scope - a lexical scope but has a place for local variables,
parameters, and type parameters.
* Named type - a member scope for named types (classes, records, enums,
etc.)
  * Namespace - a named type scope for namespaces.
  * Source file scope - a lexical scope which handles usings and etc.
* Added CSharp symbols following Roslyn's object model. CSharpSymbol is
the base symbol, then there are various subtypes when needed (e.g. to
support additional methods, properties, or declaration/member spaces).

## Changes to Java

* Added Java symbols and scopes.
* Added `LexicalDeclaration` to create Java's specific lexical
declaration scope.

## Changes to JSON

* Refactor to use symbol flow instead of custom contexts.
* Add a JSON Symbol

## Changes to Python

* Removed implicit name-based refkeys.
* Added `LexicalScope` and `MemberScope` components.
* Removed symbol flow from `ClassDeclaration` as it isn't necessary
(members are created by explicitly adding the symbols into the current
scope)
* The `Declaration` component no longer creates a member scope for
member declarations. Python will likely need different member scopes for
different cases eventually so it's probably better to be explicit.
* Some type declarations had both a member scope and regular scope
inside of it. There is now just a member scope. This means that some
things that were previously making local declarations are now making
members (e.g. static/class methods).
* Aligned name conflict resolution with the other languages by
extracting the functionality out into a function you can pass to the
binder.
* Removed the top-level python scope which set name conflict resolution
as it was no longer needed.
* Cleaned up the symbol creation function. Mostly made binder and scope
set inside this function so callers don't need to provide it.
* Reference now builds its member expression using the member expression
components rather than strings.

## Changes to TypeScript

* Added `LexicalScope and `MemberScope` components to make these
TypeScript specific scopes.
* Added type and value declaration spaces. These no longer conflict
(except in the case of classes which introduce both a type and a value).
* `Declaration` no longer makes a member scope. Explicitly create the
proper one when needed.
* Member expression supports jsPrivate.
* Removed some usage of AssignmentContext, replacing with symbol flow.
* Removed `PrivateScopeContext` as it is no longer needed.
* Added a ton of factory methods to simplify creating symbols in the
proper scope.