Add stdlib for C#, various related improvements to referencing symbols (#295)
With this PR, you can now reference types from the System namespace by
importing the appropriate namespace from `@alloy-js/csharp/global`. This
is achieved by processing docfx output for C#'s System namespace into
TypeScript definitions of the symbols. Various related improvements are
described in the changelog.

This presently doesn't build due to the sheer size of the inferred types
here, so I'm working on emitting the interface as well.