Add support for extra imports (#342)
There are certain scenarios in Python where we need to add imports at
the top of the file without them necessarily being triggered by a symbol
being referenced. One example is `from __future__ import annotations`.

To allow that, we are adding the `futureImports` prop, which already
existed and wasn't being used.

Also, we are making the module docstring to be rendered at the top, as
well as properly formatting the header and headerComment.

Finally, we are fixing some spacing-related situations to better adhere
to PEP8.

With those changes, this is how a rendered file would look:

```
<module docstring>
<manually inserted imports>
<imports generated by the import engine>
<children>
```